feat: move delete job logic to k8sjob utils
